Change sim_demo for non-Windows compatibility
Our code should run on any computer that downloads the github repo, with MINIMAL
SETUP. ALL the code we need should be in the github repository or sourced from
it.

Paths:
Please don't use absolute paths (D:\) as they are computer specific. I have
changed these paths so that the data is generated in the repository and should
work ON ALL DEVICES. Keep everything in the repository. Please.

Please don't use back slashes (\) as they only work on Windows. Forward slashes
(/) work on Windows, Mac, and Linux.

The "fullfile" matlab function can help us make paths that work on all devices,
as they generate the slashes as needed to match the needs of the computer.
Please use this.

Github commits:
"Added file via upload" is not an adequate commit message. You must explain what
your code does and why you are making changes. It should have the word "Fix"
"Add" or "Change" in the title, and explain WHY.

The point of the github is to keep track of our changes. It is not a dropbox.
This is how long a github commit message should be.